backend/requirements.txt
fastapi[standard]==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
openpyxl==3.1.5
pydantic==2.11.3
pydantic-settings==2.9.1
alembic==1.15.2
